+++ src/neuroconv/tools/audio/__init__.py
@@ -1,1 +1,2 @@
 from .audio import add_acoustic_waveform_series
+from .mp3 import read_mp3